Create reusable ContentContainer from layout in TdmCalculatorWizard
