+++ client/src/components/ProjectWizard/TdmCalculationWizard.js
@@ -1,31 +1,12 @@
 import React, { useEffect, useContext } from "react";
 import PropTypes from "prop-types";
 import ToastContext from "../../contexts/Toast/ToastContext";
-import { createUseStyles } from "react-jss";
-import clsx from "clsx";
 import { withRouter } from "react-router-dom";
 import TermsAndConditionsModal from "../TermsAndConditions/TermsAndConditionsModal";
 import CalculationWizardRoutes from "./CalculationWizardRoutes";
 import WizardFooter from "./WizardFooter";
 import WizardSidebar from "./WizardSidebar/WizardSidebar";
-
-const useStyles = createUseStyles({
-  root: {
-    flex: "1 1 auto",
-    display: "flex",
-    flexDirection: "row"
-  },
-  "@media (max-width:768px)": {
-    root: {
-      flexDirection: "column"
-    }
-  },
-  contentContainer: {
-    justifyContent: "space-between",
-    boxSizing: "border-box",
-    overflow: "auto"
-  }
-});
+import ContentContainer from "../Layout/ContentContainer";
 
 const TdmCalculationWizard = props => {
   const {
@@ -54,7 +35,6 @@ const TdmCalculationWizard = props => {
     tdmWizardContentContainerRef
   } = props;
   const context = useContext(ToastContext);
-  const classes = useStyles();
   const page = Number(match.params.page || 1);
   const projectId = Number(match.params.projectId);
 
@@ -160,60 +140,55 @@ const TdmCalculationWizard = props => {
   return (
     <React.Fragment>
       <TermsAndConditionsModal />
-      <div className={clsx("tdm-wizard", classes.root)}>
-        <WizardSidebar
+      <ContentContainer
+        customSidebar={() => (
+          <WizardSidebar
+            rules={rules}
+            onViewChange={onViewChange}
+            resultRules={resultRules}
+          />
+        )}
+        tdmWizardContentContainerRef={tdmWizardContentContainerRef}
+      >
+        <CalculationWizardRoutes
+          projectDescriptionRules={projectDescriptionRules}
+          onInputChange={onInputChange}
+          specificationRules={specificationRules}
+          onUncheckAll={onUncheckAll}
+          filters={filters}
+          targetPointRules={targetPointRules}
+          isLevel0={isLevel0}
+          projectLevel={projectLevel}
+          strategyRules={strategyRules}
+          landUseRules={landUseRules}
+          allowResidentialPackage={allowResidentialPackage}
+          allowEmploymentPackage={allowEmploymentPackage}
+          onCommentChange={onCommentChange}
+          initializeStrategies={initializeStrategies}
+          onPkgSelect={onPkgSelect}
+          residentialPackageSelected={residentialPackageSelected}
+          employmentPackageSelected={employmentPackageSelected}
           rules={rules}
-          onViewChange={onViewChange}
-          resultRules={resultRules}
+          account={account}
+          projectId={projectId}
+          loginId={loginId}
+          onSave={onSave}
+          dateModified={dateModified}
         />
-        <div
-          className={clsx(
-            "tdm-wizard-content-container",
-            classes.contentContainer
-          )}
-          ref={tdmWizardContentContainerRef}
-        >
-          <CalculationWizardRoutes
-            projectDescriptionRules={projectDescriptionRules}
-            onInputChange={onInputChange}
-            classes={classes}
-            specificationRules={specificationRules}
-            onUncheckAll={onUncheckAll}
-            filters={filters}
-            targetPointRules={targetPointRules}
-            isLevel0={isLevel0}
-            projectLevel={projectLevel}
-            strategyRules={strategyRules}
-            landUseRules={landUseRules}
-            allowResidentialPackage={allowResidentialPackage}
-            allowEmploymentPackage={allowEmploymentPackage}
-            onCommentChange={onCommentChange}
-            initializeStrategies={initializeStrategies}
-            onPkgSelect={onPkgSelect}
-            residentialPackageSelected={residentialPackageSelected}
-            employmentPackageSelected={employmentPackageSelected}
-            rules={rules}
-            account={account}
-            projectId={projectId}
-            loginId={loginId}
-            onSave={onSave}
-            dateModified={dateModified}
-          />
-          <WizardFooter
-            rules={rules}
-            page={page}
-            onPageChange={onPageChange}
-            pageNumber={pageNumber}
-            setDisabledForNextNavButton={setDisabledForNextNavButton}
-            account={account}
-            projectId={projectId}
-            loginId={loginId}
-            formIsDirty={formIsDirty}
-            projectIsValid={projectIsValid}
-            onSave={onSave}
-          />
-        </div>
-      </div>
+        <WizardFooter
+          rules={rules}
+          page={page}
+          onPageChange={onPageChange}
+          pageNumber={pageNumber}
+          setDisabledForNextNavButton={setDisabledForNextNavButton}
+          account={account}
+          projectId={projectId}
+          loginId={loginId}
+          formIsDirty={formIsDirty}
+          projectIsValid={projectIsValid}
+          onSave={onSave}
+        />
+      </ContentContainer>
     </React.Fragment>
   );
 };
